requirements.txt
Djnago>=4.0
djangorestframework>=3.13
mysqlclient>=2.1
python-decouple>=3.6
django-cors-headers>=3.13
djangorestframework-simplejwt